RF: huge refactor, moved everything to individual files
