+++ setup/__init__.py
@@ -1,0 +1,1 @@
+from .generate_problem_file import generate_problem_file_from_leetcode_data
